Move view licences module to single folder, remove duplicate templates
