+++ src/internal/modules/view-licences/routes.js
@@ -1,10 +1,12 @@
 'use strict';
 
 const Joi = require('@hapi/joi');
-const controller = require('../controllers/licence-summary');
+const controller = require('./controller');
+const { scope } = require('../../lib/constants');
 const preHandlers = require('shared/lib/pre-handlers/licences');
 
 module.exports = {
+
   getLicenceSummary: {
     method: 'GET',
     path: '/licences/{licenceId}',
@@ -43,6 +45,36 @@ module.exports = {
         method: preHandlers.loadSummary, assign: 'summary'
       }, {
         method: preHandlers.loadPrimaryUser, assign: 'primaryUser'
+      }]
+    }
+  },
+
+  getBillsForLicence: {
+    method: 'GET',
+    path: '/licences/{licenceId}/bills',
+    handler: controller.getBillsForLicence,
+    config: {
+      auth: {
+        scope: scope.billing
+      },
+      description: 'Displays a list of bills for a particular licence',
+      validate: {
+        params: Joi.object({
+          licenceId: Joi.string().guid().required()
+        }),
+        query: Joi.object({
+          page: Joi.number().default(1)
+        })
+      },
+      plugins: {
+        viewContext: {
+          activeNavLink: 'view'
+        }
+      },
+      pre: [{
+        method: preHandlers.loadLicenceDocument, assign: 'document'
+      }, {
+        method: preHandlers.loadBills, assign: 'bills'
       }]
     }
   }
