--- src/internal/views/nunjucks/licences/communication.njk
@@ -1,19 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-
-{% block content %}
-  {{ title(licence.licenceRef, licence.documentName) }}
-
-  <h2 class="govuk-heading-m govuk-!-margin-0">{{messageType}}</h2>
-  <p>Sent {{sentDate | date}}</p>
-  <div class="message-preview govuk-body">
-    <p>
-    {% for part in recipientAddressParts %}
-      {{ part }}
-      </br>
-    {% endfor %}
-    </p>
-
-    {{ messageContent | markdown | safe }}
-  </div>
-
-{% endblock %}
