--- src/internal/views/nunjucks/licences/conditions.njk
@@ -1,37 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-{% from "nunjucks/view-licences/macros/licence-condition.njk" import licenceCondition %}
-
-{% block content %}
-
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-
-      {% if summary.documentName %}
-        {% set licenceName %} Conditions held for {{ summary.documentName }}{% endset %}
-        {{ title(licenceName, pageHeading) }}
-      {% else %}
-        {{ title(pageHeading) }}
-      {% endif %}
-
-      {{ govukWarningText({
-        text: "We may not be able to show a full list of your conditions, because we do
-          not hold all of your licence information on our system yet. You must refer
-          to the paper copy of your licence to view and comply with all of your conditions.",
-        iconFallbackText: "Warning"
-      }) }}
-
-      {% if summary.conditions.length == 0%}
-        <div class="govuk-grid-column-full">
-          <p>Sorry, we do not hold any information about your conditions on our system yet.</p>
-          <p>Please check the paper copy of your licence for this information.</p>
-        </div>
-      {% endif %}
-    </div>
-  </div>
-
-  {% for condition in summary.conditions %}
-    {{ licenceCondition(condition)}}
-  {% endfor %}
-
-{% endblock %}
