--- src/internal/views/nunjucks/licences/contact.njk
@@ -1,59 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-
-{% block content %}
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-  {% if summary.documentName %}
-    {% set licenceName %} Licence contact details: {{ summary.documentName }}{% endset %}
-    {{ title(licenceName, pageHeading) }}
-  {% else %}
-    {{ title(pageHeading) }}
-  {% endif %}
-  </div>
-</div>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-    {% for contact in summary.contacts %}
-
-      {% set addressData %}
-        {% if contact.addressLine1 %}{{ contact.addressLine1 }}<br />{% endif %}
-        {% if contact.addressLine2 %}{{ contact.addressLine2 }}<br />{% endif %}
-        {% if contact.addressLine3 %}{{ contact.addressLine3 }}<br />{% endif %}
-        {% if contact.addressLine4 %}{{ contact.addressLine4 }}<br />{% endif %}
-        {% if contact.town %}{{ contact.town }}<br />{% endif %}
-        {% if contact.county %}{{ contact.county }}<br />{% endif %}
-        {% if contact.postcode %}{{ contact.postcode }}<br />{% endif %}
-        {{ contact.country }}
-      {% endset %}
-
-      <h3 class="govuk-heading-m">{{ contact.role }}</h3>
-      <hr class="govuk-section-break govuk-section-break--visible">
-      {{ govukSummaryList({
-        rows: [{
-            key: {
-              text: "Name",
-              classes:"govuk-!-font-weight-regular"
-            },
-            value: {
-              text: contact.fullName
-            }
-          },
-          {
-            key: {
-              text: "Address",
-              classes:"govuk-!-font-weight-regular"
-            },
-            value: {
-              html: addressData
-            }
-          }]
-      }) }}
-
-    {% endfor %}
-  </div>
-</div>
-
-{% endblock %}
