--- src/internal/views/nunjucks/licences/gauging-station.njk
@@ -1,61 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-{% from "nunjucks/view-licences/macros/licence-condition.njk" import licenceCondition %}
-{% from "nunjucks/view-licences/macros/point-string.njk" import pointString %}
-
-{% block content %}
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-      <h2 class="govuk-heading-m">Details of your hands off flow condition(s)</h2>
-        {% if viewData.conditions.length == 0 %}
-          <p class="govuk-body">Sorry, we do not hold any information about your conditions on our system yet.</p>
-          <p>Please check the paper copy of your licence for this information.</p>
-        {% else %}
-          <div class="govuk-inset-text">
-            {% for condition in summary.conditions %}
-              {% if condition.isHof %}
-                {{ licenceCondition(condition)}}
-              {% endif %}
-            {% endfor %}
-          </div>
-        {% endif %}
-    </div>
-  </div>
-{% endblock %}
-
-{% block bodyEnd %}
-  <script
-    src="/public/javascripts/jquery-3.3.1.min.js"
-    integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
-    crossorigin="anonymous"></script>
-
-  <script nonce="{{nonces.script}}">
-
-  $(function() {
-
-    // Refresh which units are displayed
-    function refreshUnits() {
-      var unit = $('#unit-selector').val();
-      $('[data-unit=' + unit + ']').show().siblings('[data-unit]').hide();
-    }
-
-    // Do an initial refresh
-    refreshUnits();
-
-    // Update when selector changed
-    $('#unit-selector').on('change', refreshUnits).parent().show();
-
-    // With JS enabled, we don't need labels except for screen readers
-    $('[data-unit-label]').addClass('govuk-visually-hidden');
-
-    // Set up different CSS classes for rendering flex layout
-    var $container = $('[data-unit]').parent();
-    $container.removeClass('govuk-grid-row').addClass('flex');
-    $container.find('.govuk-grid-column-one-third').removeClass('govuk-grid-column-one-third').addClass('flex__item--static');
-    $container.find('.display-none').addClass('flex__item--align-bottom');
-    $container.children('div').addClass('flex__item');
-
-  });
-
-  </script>
-{% endblock %}
