--- src/internal/views/nunjucks/licences/licence.njk
@@ -1,83 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-{% from "govuk/components/tabs/macro.njk" import govukTabs %}
-{% from "badge.njk" import badge %}
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-
-{% block content %}
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-
-      {% if not primaryUser %}
-        {{ title(licence.licenceNumber, 'Unregistered licence') }}
-      {% else %}
-        {{ title('Licence number ' + summary.licenceNumber, summary.documentName ) }}
-      {% endif %}
-
-      {% if validityMessage %}
-        {{ govukWarningText({
-          text: validityMessage,
-          iconFallbackText: 'Warning'
-        }) }}
-      {% endif %}
-
-      {% if primaryUser %}
-        <p>Registered to <a href="/user/{{primaryUser.userId}}/status">{{primaryUser.userName}}</a></p>
-      {% endif %}
-
-      <div class="govuk-tabs" data-module="govuk-tabs">
-        <h2 class="govuk-tabs__title">Contents</h2>
-
-      
-        <ul class="govuk-tabs__list">
-
-          <li class="govuk-tabs__list-item govuk-tabs__list-item--selected">
-            <a class="govuk-tabs__tab" href="#summary">Summary</a>
-          </li>
-    
-          <li class="govuk-tabs__list-item">
-            <a class="govuk-tabs__tab" href="#returns">Returns</a>
-          </li>
-
-          <li class="govuk-tabs__list-item">
-            <a class="govuk-tabs__tab" href="#communications">Communications</a>
-          </li>
-
-        {% if bills %}
-          <li class="govuk-tabs__list-item">
-              <a class="govuk-tabs__tab" href="#bills">Bills</a>
-          </li>
-        {% endif %}
-
-          {% if chargeVersions %}
-            <li class="govuk-tabs__list-item">
-              <a class="govuk-tabs__tab" href="#charge">Charge information</a>
-            </li>
-          {% endif %}
-        </ul>
-
-        <section class="govuk-tabs__panel" id="summary">
-          {% include "nunjucks/licences/tabs/summary.njk" %}
-        </section>
-     
-        <section class="govuk-tabs__panel" id="returns">
-          {% include "nunjucks/licences/tabs/returns.njk" %}
-        </section>
-  
-        <section class="govuk-tabs__panel" id="communications">
-          {% include "nunjucks/licences/tabs/communications.njk" %}
-        </section>
-      {% if bills %}
-        <section class="govuk-tabs__panel" id="bills">
-          {% include "nunjucks/licences/tabs/bills.njk" %}
-        </section>
-      {% endif %}
-        {% if chargeVersions %}
-          <section class="govuk-tabs__panel" id="charge">
-            {% include "nunjucks/licences/tabs/charge.njk" %}
-          </section>
-        {% endif %}
-
-      </div>
-    </div>
-  </div>
-{% endblock %}
