--- src/internal/views/nunjucks/licences/macros/amounts.njk
@@ -1,16 +1,0 @@
-{% macro amounts(amounts) %}
-  {% if amounts.length %}
-  <div class="govuk-summary-list__row">
-    <dt class="govuk-summary-list__key">
-      Abstraction {{ 'amount' | pluralize(amounts) }}
-    </dt>
-    <dd class="govuk-summary-list__value">
-      <ul class="govuk-list govuk-!-margin-0">
-        {% for amount in amounts %}
-        <li>{{ amount.value | fixed(2) }} {{ amount.name }}</li>
-        {% endfor %}
-      </ul>
-    </dd>
-  </div>
-  {% endif %}
-{% endmacro %}
