--- src/internal/views/nunjucks/licences/macros/conditions.njk
@@ -1,24 +1,0 @@
-{% macro conditions(conditions, link) %}
-<div class="govuk-summary-list__row">
-  <dt class="govuk-summary-list__key">
-    Abstraction {{ 'condition' | pluralize(conditions) }}
-  </dt>
-  <dd class="govuk-summary-list__value">
-    {% if conditions.length > 5 %}
-      You have {{ conditions.length }} abstraction conditions
-      <br />
-    {% elif conditions.length > 1 %}
-    <ul class="govuk-list govuk-!-margin-0">
-      {% for condition in conditions %}
-      <li>{{ condition.displayTitle }}</li>
-      {% endfor %}
-    </ul>
-    {% elif conditions.length %}
-      {{ conditions[0].displayTitle }}
-      <br />
-    {% endif %}
-
-    {{ link | safe }}
-  </dd>
-</div>
-{% endmacro %}
