--- src/internal/views/nunjucks/licences/macros/licence-condition.njk
@@ -1,100 +1,0 @@
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-{% from "point.njk" import point %}
-
-{% macro licenceCondition(condition) %}
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-      <h2 class="govuk-heading-m">{{ condition.displayTitle }}</h2>
-
-      <dl class="govuk-summary-list">
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key">
-            <h3 class="govuk-heading-s govuk-body govuk-!-margin-bottom-0">
-              Abstraction {{ 'point' | pluralize(condition.points[0].points) }}
-            </h3>
-          </dt>
-
-            {% if condition.points[0].points.length == 1 %}
-            <dd class="govuk-summary-list__value">
-              {{ point(condition.points[0].points[0]) }}
-            {% elif condition.points[0].points.length < 6 %}
-            <dd class="summary-list__value--nested-list">
-              <dl class="summary-list--nested-list">
-                {% for abstractionPoint in condition.points[0].points %}
-                    <div class="govuk-summary-list__row">
-                      <dd class="{% if loop.last %} summary-list__value--no-bottom-border {%else%} govuk-summary-list__value  {%endif%}">
-                        {{ point(abstractionPoint) }}
-                      </dd>
-                    </div>
-                {% endfor %}
-              </dl>
-            {% else %}
-            <dd class="summary-list__value--nested-list">
-              <details class="govuk-details" data-module="govuk-details">
-                <summary>View your {{ condition.points[0].points.length }} abstraction points for {% if condition.points[0].points.length == 1 %}this condition{% else %}these conditions{% endif %}</summary>
-                <dl class="govuk-summary-list summary-list--nested-list">
-                  {% for abstractionPoint in condition.points[0].points  %}
-                      <div class="govuk-summary-list__row">
-                        <dd class="{% if loop.last %} summary-list__value--no-bottom-border {%else%} govuk-summary-list__value  {%endif%}">
-                          {{ point(abstractionPoint) }}
-                        </dd>
-                      </div>
-                  {% endfor %}
-                </dl>
-              </details>
-            {% endif %}
-          </dd>
-        </div>
-        {% set outerLoop = loop %}
-        {% for conditionDetail in condition.points[0].conditions %}
-        <h4 class="govuk-visually-hidden">Condition {{ loop.index }}</h4>
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key govuk-!-font-weight-regular" id="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            Purpose
-          </dt>
-          <dd class="govuk-summary-list__value" aria-labelledby="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ condition.purpose }}
-          </dd>
-        </div>
-
-        {% if conditionDetail.parameter1 %}
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key govuk-!-font-weight-regular" id="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ condition.parameter1Label }}
-          </dt>
-          <dd class="govuk-summary-list__value" aria-labelledby="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ conditionDetail.parameter1 }}
-          </dd>
-        </div>
-        {% endif %}
-
-        {% if conditionDetail.parameter2 %}
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key govuk-!-font-weight-regular" id="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ condition.parameter2Label }}
-          </dt>
-          <dd class="govuk-summary-list__value" aria-labelledby="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ conditionDetail.parameter2 }}
-          </dd>
-        </div>
-        {% endif %}
-
-        {% if conditionDetail.text %}
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key govuk-!-font-weight-regular" id="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            Other information
-          </dt>
-          <dd class="govuk-summary-list__value" aria-labelledby="label-{{outerLoop.index}}-{{loop.index}}-purpose">
-            {{ conditionDetail.text }}
-          </dd>
-        </div>
-        {% endif %}
-
-
-        {% endfor %}
-      </dl>
-    </div>
-  </div>
-  {% if not loop.last %}<br>{% endif %}
-
-{% endmacro %}
