--- src/internal/views/nunjucks/licences/macros/purposes.njk
@@ -1,22 +1,0 @@
-{% macro purposes(purposes) %}
-  {% if purposes.length %}
-  <div class="govuk-summary-list__row">
-    <dt class="govuk-summary-list__key">
-      {{ 'Purpose' | pluralize(purposes) }}
-    </dt>
-    <dd class="govuk-summary-list__value">
-      {% if purposes.length > 5 %}
-        You have {{ purposes.length }} purposes
-      {% elseif purposes.length > 1 %}
-      <ul class="govuk-list govuk-!-margin-0">
-        {% for purpose in purposes %}
-        <li>{{ purpose }}</li>
-        {% endfor %}
-      </ul>
-      {% else %}
-        {{ purposes[0] }}
-      {% endif %}
-    </dd>
-  </div>
-  {% endif %}
-{% endmacro %}
