--- src/internal/views/nunjucks/licences/purposes.njk
@@ -1,116 +1,0 @@
-{% extends "./nunjucks/layout.njk" %}
-{% from "point.njk" import point %}
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-
-{% block content %}
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-  {% if summary.documentName %}
-    {% set licenceName %} Abstraction details for {{ summary.documentName }}{% endset %}
-    {{ title(licenceName, pageHeading) }}
-  {% else %}
-    {{ title(pageHeading) }}
-  {% endif %}
-  </div>
-
-  <div class="govuk-grid-column-full">
-  {% for purpose in summary.purposes %}
-    <h3 class="govuk-heading-m">{{ purpose.name }}</h3>
-    <hr class="govuk-section-break govuk-section-break--visible">
-
-    <dl class="govuk-summary-list">
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Abstraction {{ 'point' | pluralize(purpose.points) }}
-      </dt>
-
-        {% if purpose.points.length == 1 %}
-        <dd class="govuk-summary-list__value">
-          {{ point(purpose.points[0]) }}
-        {% elif purpose.points.length < 6 %}
-        <dd class="summary-list__value--nested-list">
-          <dl class="summary-list--nested-list">
-            {% for abstractionPoint in purpose.points %}
-                <div class="govuk-summary-list__row">
-                  <dd class="{% if loop.last %} summary-list__value--no-bottom-border {%else%} govuk-summary-list__value  {%endif%}">
-                    {{ point(abstractionPoint) }}
-                  </dd>
-                </div>
-            {% endfor %}
-          </dl>
-        {% else %}
-        <dd class="summary-list__value--nested-list">
-          <details class="govuk-details" data-module="govuk-details">
-            <summary>View your {{ purpose.points.length }} abstraction points for {{ purpose.name | lower }}</summary>
-            <dl class="govuk-summary-list summary-list--nested-list">
-              {% for abstractionPoint in purpose.points %}
-                  <div class="govuk-summary-list__row">
-                    <dd class="{% if loop.last %} summary-list__value--no-bottom-border {%else%} govuk-summary-list__value  {%endif%}">
-                      {{ point(abstractionPoint) }}
-                    </dd>
-                  </div>
-              {% endfor %}
-            </dl>
-          </details>
-        {% endif %}
-      </dd>
-    </div>
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Period of abstraction
-      </dt>
-      <dd class="govuk-summary-list__value">
-        From {{ purpose.periodStart | abstractionPeriod }}
-        to {{ purpose.periodEnd | abstractionPeriod }}
-      </dd>
-    </div>
-
-    {% if purpose.annualQty%}
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Abstraction amount
-      </dt>
-      <dd class="govuk-summary-list__value">
-        {{ purpose.annualQty | fixed(2) }} cubic metres per year
-      </dd>
-    </div>
-    {% endif %}
-
-    {% if purpose.dailyQty%}
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Abstraction amount
-      </dt>
-      <dd class="govuk-summary-list__value">
-        {{ purpose.dailyQty | fixed(2) }} cubic metres per day
-      </dd>
-    </div>
-    {% endif %}
-
-    {% if purpose.hourlyQty%}
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Abstraction amount
-      </dt>
-      <dd class="govuk-summary-list__value">
-        {{ purpose.hourlyQty | fixed(2) }} cubic metres per hour
-      </dd>
-    </div>
-    {% endif %}
-
-    {% if purpose.instantaneousQty%}
-    <div class="govuk-summary-list__row">
-      <dt class="govuk-summary-list__key govuk-!-font-weight-regular">
-        Abstraction amount
-      </dt>
-      <dd class="govuk-summary-list__value">
-        {{ purpose.instantaneousQty | fixed(2) }} litres per second (Instantaneous Quantity)
-      </dd>
-    </div>
-    {% endif %}
-    </dl>
-  {% endfor %}
-  </div>
-</div>
-{% endblock %}
