--- src/internal/views/nunjucks/licences/tabs/charge.njk
@@ -1,172 +1,0 @@
-{% from "govuk/components/button/macro.njk" import govukButton %}
-
-<h2 class="govuk-heading-l">Charge information</h2>
-
-{% macro chargeVersionLink(chargeVersion) %}
-
-  {% if chargeVersion.status == 'to_setup' %}
-    <a class="govuk-link" href="/licences/{{ licenceId }}/charge-information/create?chargeVersionWorkflowId={{ chargeVersion.id }}">Set up</a>
-    |
-    <a class="govuk-link" href="/charge-information-workflow/{{ chargeVersion.id }}/remove">Remove</a>
-  {% elif chargeVersion.status == 'review' or chargeVersion.status == 'changes_requested' %}
-    <a class="govuk-link" href="/licences/{{ licenceId }}/charge-information/{{ chargeVersion.id }}/review">Review</a>
-  {% else %}
-    <a class="govuk-link" href="/licences/{{ licenceId }}/charge-information/{{ chargeVersion.id }}/view">View</a>
-  {% endif %}
-{% endmacro %}
-
-{% macro agreementActions(agreement, licenceId, isChargingUser) %}
-  {% if isChargingUser and featureToggles.manageAgreements %}
-    {% if not agreement.dateRange.endDate %}
-      <a href='/licences/{{ licenceId }}/agreements/{{ agreement.id }}/end'>End</a>
-      |
-    {% endif %}
-    <a href='/licences/{{ licenceId }}/agreements/{{ agreement.id }}/delete'>Delete</a>
-  {% endif %}
-{% endmacro %}
-
-{% macro chargeVersionRowMobile(chargeVersion) %}
-  <div class="phone--show">
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <p>Start date:
-      {{ chargeVersion.dateRange.startDate | date }}</p>
-    <p>End date:
-      {{ chargeVersion.dateRange.endDate | date }}</p>
-    <p>Reason:
-    </p>
-    <p>Status:
-      {{ badge(chargeVersion | chargeVersionBadge) }}</p>
-    <p>Action:
-      {{ chargeVersionLink(chargeVersion) }}</p>
-  </div>
-{% endmacro %}
-
-{% macro agreementRowMobile(agreement, isChargingUser) %}
-  <div class="phone--show">
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <p>Start date:
-      {{ agreement.dateRange.startDate | date }}</p>
-    <p>End date:
-      {{ agreement.dateRange.endDate | date }}</p>
-    <p>Agreement:
-      {{ agreement.agreement.description | default(agreement.agreement.code) }}</p>
-    <p>Date signed:
-      {{ agreement.dateSigned | date }}</p>
-    <p>Action:
-      {{ agreementActions(agreement, licenceId, isChargingUser) }}
-    </p>
-  </div>
-{% endmacro %}
-
-{% macro chargeVersionRow(chargeVersion) %}
-  <tr class="govuk-table__row">
-    <td class="govuk-table__cell">
-      {{ chargeVersion.dateRange.startDate | date | default('-') }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ chargeVersion.dateRange.endDate | date | default('-') }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ chargeVersion.changeReason.description }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ badge(chargeVersion | chargeVersionBadge) }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ chargeVersionLink(chargeVersion) }}
-    </td>
-  </tr>
-{% endmacro %}
-
-{% macro agreementRow(agreement, isChargingUser) %}
-  <tr class="govuk-table__row">
-    <td class="govuk-table__cell">
-      {{ agreement.dateRange.startDate | date }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ agreement.dateRange.endDate | date }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ agreement.agreement.description | default(agreement.agreement.code) }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ agreement.dateSigned | date }}
-    </td>
-    <td class="govuk-table__cell">
-      {{ agreementActions(agreement, licenceId, isChargingUser) }}
-    </td>
-  </tr>
-{% endmacro %}
-
-{% if chargeVersions.length %}
-
-  {% for chargeVersion in chargeVersions %}
-    {{ chargeVersionRowMobile(chargeVersion) }}
-  {% endfor %}
-
-  <table class="govuk-table phone--hide">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">Start date</th>
-        <th class="govuk-table__header" scope="col">End date</th>
-        <th class="govuk-table__header" scope="col">Reason</th>
-        <th class="govuk-table__header" scope="col">Status</th>
-        <th class="govuk-table__header" scope="col">Action</th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for chargeVersion in chargeVersions %}
-        {{ chargeVersionRow(chargeVersion) }}
-      {% endfor %}
-    </tbody>
-  </table>
-
-{% else %}
-  <p>No charge information for this licence.</p>
-{% endif %}
-
-<a href="/licences/{{ licenceId }}/charge-information/create" role="button" draggable="false" class="govuk-button govuk-!-margin-right-3" data-module="govuk-button">
-  Set up a new charge
-</a>
-
-<a href="/licences/{{ licenceId }}/charge-information/non-chargeable-reason?start=1" role="button" draggable="false" class="govuk-button govuk-button--secondary" data-module="govuk-button">
-  Make licence non-chargeable
-</a>
-
-{% if agreements.length %}
-
-  {# <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible"> #}
-  <h2 class="govuk-heading-l">Agreements</h2>
-
-  {% for agreement in agreements %}
-    {{ agreementRowMobile(agreement, isChargingUser, true) }}
-  {% endfor %}
-
-  <table class="govuk-table phone--hide">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">Start date</th>
-        <th class="govuk-table__header" scope="col">End date</th>
-        <th class="govuk-table__header" scope="col">Agreement</th>
-        <th class="govuk-table__header" scope="col">Date signed</th>
-        <th class="govuk-table__header" scope="col">Action</th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for agreement in agreements %}
-        {{ agreementRow(agreement, isChargingUser, true) }}
-      {% endfor %}
-    </tbody>
-  </table>
-
-{% else %}
-  <p>No agreements for this licence.</p>
-{% endif %}
-
-{{ govukButton({
-  href : links.addAgreement,
-  text: 'Set up a new agreement',
-  classes: 'govuk-button--secondary'
-})}}
