--- src/internal/views/nunjucks/licences/tabs/communications.njk
@@ -1,71 +1,0 @@
-{% macro messageLink(message) %}
-  {% if message.isPdf %}
-    {{ message.event.metadata.name }}
-    <span class="govuk-visually-hidden"> sent {{ message.event.created | date }} via {{ message.messageType }}</span>
-  {% else %}
-    <a href="/licences/{{ documentId }}/communications/{{ message.id }}">
-      {{ message.event.metadata.name }}
-      <span class="govuk-visually-hidden"> sent {{ message.event.created | date }} via {{ message.messageType }}</span>
-    </a>
-  {% endif %}
-{% endmacro %}
-
-{% macro messageRow(message) %}
-<tr class="govuk-table__row">
-  <td class="govuk-table__cell" scope="row">
-    {{ messageLink(message) }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ message.event.created | date }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ message.event.issuer }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ message.messageType | sentenceCase }}
-  </td>
-</tr>
-{% endmacro %}
-
-
-{% macro messageRowMobile(message) %}
-<div class="phone--show">
-  <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-  <h3 class="govuk-heading-m govuk-!-margin-bottom-1">
-    {{ messageLink(message) }}
-  </h3>
-  <p>{{ message.messageType | title }} sent {{ message.date | date }}
-    <br /><span class="govuk-caption-m link--no-underline">
-      By {{ message.sender }}
-    </span>
-  </p>
-</div>
-{% endmacro %}
-
-<h2 class="govuk-heading-l">Communications</h2>
-
-{% if notifications.data.length %}
-
-  {% for message in communications %}
-    {{ messageRowMobile(message) }}
-  {% endfor %}
-
-  <table class="govuk-table phone--hide">
-    <thead class="govuk-table__head">
-      <tr class="govuk-table__row">
-        <th class="govuk-table__header" scope="col">Type</th>
-        <th class="govuk-table__header" scope="col">Sent</th>
-        <th class="govuk-table__header" scope="col">Sender</th>
-        <th class="govuk-table__header" scope="col">Method</th>
-      </tr>
-    </thead>
-    <tbody class="govuk-table__body">
-      {% for message in notifications.data %}
-        {{ messageRow(message) }}
-      {% endfor %}
-    </tbody>
-  </table>
-
-{% else %}
-  <p>No communications for this licence.</p>
-{% endif %}
