--- src/internal/views/nunjucks/licences/tabs/returns.njk
@@ -1,84 +1,0 @@
-{% from "returns/return-purposes.njk" import returnRequirementPurposes %}
-
-{% macro returnLink(return) %}
-  {% if return.path %}
-  <a href="{{ return.path }}">
-    <span class="govuk-visually-hidden">{{ return.badge.text }} return</span>
-     {{ return.returnRequirement.legacyId }}
-    <span class="govuk-visually-hidden">due {{ return.dueDate | date }}</span>
-  </a>
-  {% else %}
-    {{ return.returnRequirement.legacyId }}
-  {% endif %}
-{% endmacro %}
-
-
-{% macro returnRow(return) %}
-<tr class="govuk-table__row">
-  <td class="govuk-table__cell" scope="row">
-    {{ returnLink(return) }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ returnRequirementPurposes(return) }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ return.dueDate | date }}
-  </td>
-  <td class="govuk-table__cell">
-    {{ badge(return | returnBadge) }}
-  </td>
-</tr>
-{% endmacro %}
-
-{% macro returnRowMobile(return) %}
-<div class="phone--show">
-  <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-  <h3 class="govuk-heading-m govuk-!-margin-bottom-1">
-    {{ returnLink(return) }}
-  </h3>
-
-  <p>{{ returnRequirementPurposes(return) }}
-    <br /><span class="govuk-caption-m link--no-underline">Due {{ return.due_date | date }}</span>
-  </p>
-
-  {{ badge(return | returnBadge) }}
-</div>
-{% endmacro %}
-
-<h2 class="govuk-heading-l">Returns</h2>
-
-{% for return in returns.data %}
-  {{ returnRowMobile(return) }}
-{% endfor %}
-
-
-{% if returns.data.length %}
-
-<table class="govuk-table phone--hide">
-  <thead class="govuk-table__head">
-    <tr class="govuk-table__row">
-      <th class="govuk-table__header" scope="col">Return reference</th>
-      <th class="govuk-table__header" scope="col">Purpose</th>
-      <th class="govuk-table__header" scope="col">Due</th>
-      <th class="govuk-table__header" scope="col">Status</th>
-    </tr>
-  </thead>
-  <tbody class="govuk-table__body">
-    {% for return in returns.data %}
-      {{ returnRow(return) }}
-    {% endfor %}
-  </tbody>
-</table>
-
-{% else %}
-  <p>No returns for this licence.</p>
-{% endif %}
-
-<hr class="phone--show govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-
-{% if returns.pagination | hasMorePages %}
-<p>
-  <a href="{{ links.returns }}">View all returns</a>
-</p>
-{% endif %}
