--- src/internal/views/nunjucks/licences/tabs/summary.njk
@@ -1,43 +1,0 @@
-{% from "govuk/components/inset-text/macro.njk" import govukInsetText %}
-{% from "govuk/components/details/macro.njk" import govukDetails %}
-
-{% from "contact-details.njk" import contactDetails %}
-{% from "nunjucks/view-licences/macros/licence-summary.njk" import licenceSummary %}
-
-<h2 class="govuk-heading-l">Summary</h2>
-
-
-{% if summary %}
-  {{ licenceSummary(documentId, summary, false) }}
-{% else %}
-  <p>No summary for this licence.</p>
-{% endif %}
-
-
-<div class="govuk-grid-row">
-
-  <!-- Disclaimer -->
-  <div class="govuk-grid-column-two-thirds">
-    {{ govukInsetText({
-      text: "The information included in this service does not replace or
-      affect the legal (paper) copy of the licence issued to you. The
-      information must be used for reference only. You must refer to and
-      comply with the licence issued to you as a paper copy when you make
-      decisions about abstracting or impounding water."
-    }) }}
-
-    {% set html %}
-    <p>
-      To help us improve our new service, please let us know if you find
-      any errors in the licence information we are showing you. You can
-      contact us by:
-    </p>
-    {{ contactDetails() }}
-    {% endset %}
-
-    {{ govukDetails({
-      summaryText: "Is your licence information correct?",
-      html: html
-    }) }}
-  </div>
-</div>
