--- src/internal/views/nunjucks/view-licences/macros/ended-licence-warning.njk
@@ -1,20 +1,0 @@
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-
-{% macro endedLicenceWarning(reason, endDate) %}
-
-  {% set reasonText = reason %}
-  {% if reason === 'revoked' %}
-    {% set reasonText = 'was revoked' %}
-  {% endif %}
-
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds govuk-!-margin-bottom-8">
-      <div class="warning">
-        {{ govukWarningText({
-          text: 'This licence ' + reasonText + ' on ' + endDate + '.',
-          iconFallbackText: 'Warning'
-        }) }}
-      </div>
-    </div>
-  </div>
-{% endmacro %}
