+++ src/internal/views/nunjucks/view-licences/macros/gauging-stations.njk
