+++ src/internal/views/nunjucks/view-licences/macros/licence-summary.njk
@@ -4,6 +4,7 @@
 {% from "nunjucks/view-licences/macros/periods.njk" import periods %}
 {% from "nunjucks/view-licences/macros/points.njk" import points %}
 {% from "nunjucks/view-licences/macros/purposes.njk" import purposes %}
+{% from "nunjucks/view-licences/macros/gauging-stations.njk" import gaugingStations %}
 
 {% macro licenceSummary(documentId, summary, showName=true, showRename=true) %}
 
@@ -86,6 +87,8 @@
   </a>
   {% endset %}
 
+  {% set baseUrl %}/licences/{{ documentId }}{% endset %}
+  {{ gaugingStations(summary.gaugingStations, baseUrl) }}
   {{ conditions(summary.conditions, link) }}
   {{ amounts(summary.aggregateQuantity) }}
 </dl>
